update_26_05

+++ UI/code/Login_RauMa/DAO/SanPham.cs
@@ -26,7 +26,7 @@ namespace DAO
         public Nullable<decimal> GiaTien { get; set; }
         public string MoTa { get; set; }
         public string Hinh { get; set; }
-        public Nullable<byte> TrangThai { get; set; }
+        public Nullable<short> TrangThai { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<CTHoaDon> CTHoaDons { get; set; }
